💥 breaking: adding handler capability & structure
