+++ pyproject.toml
@@ -32,6 +32,7 @@ prettytable = "^3.8.0"
 tradingview_ta = "^3.3.0"
 yfinance = "^0.2.25"
 finnhub-python = "2.4.20"
+cot_reports = "0.1.3"
 
 
 [tool.poetry.group.dev.dependencies]
